Experiments/pyproject.toml
[project]
name = "experiments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "google-cloud-vision>=3.10.2",
    "google-genai==1.11.0",
    "pandas>=2.3.2",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
]
